+++ src/Elements/ReaderElement.php
@@ -178,7 +178,7 @@ class ReaderElement extends ContentElement
         }
 
         if ($this->ad['specifics_first_models_production_date']) {
-            $this->ad['specifics_first_models_production_date'] = date($GLOBALS['TL_CONFIG']['dateFormat'], $this->ad['specifics_first_models_production_date']);
+            $this->ad['specifics_first_models_production_date'] = ListingElement::formatDate($this->ad['specifics_first_models_production_date']);
         }
 
         if ($this->ad['specifics_power']) {
@@ -275,6 +275,55 @@ class ReaderElement extends ContentElement
             $fuelConsumption[] = [
                 'label' => $GLOBALS['TL_LANG'][$this->strTable]['emission_fuel_consumption_combined_power_consumption'][0],
                 'value' => System::getFormattedNumber($this->ad['emission_fuel_consumption_combined_power_consumption'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_co2_emission_combined']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_co2_emission_combined'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_co2_emission_combined'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_consumption_fuel_combined']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_consumption_fuel_combined'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_consumption_fuel_combined'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_consumption_power_combined']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_consumption_power_combined'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_consumption_power_combined'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_electric_range']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_electric_range'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_electric_range'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_consumption_fuel_combined_weighted']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_consumption_fuel_combined_weighted'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_consumption_fuel_combined_weighted'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_consumption_power_combined_weighted']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_consumption_power_combined_weighted'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_consumption_power_combined_weighted'], 1),
+            ];
+        }
+
+        if ($this->ad['wltp_co2_emission_combined_weighted']) {
+            $fuelConsumption[] = [
+                'label' => $GLOBALS['TL_LANG'][$this->strTable]['wltp_co2_emission_combined_weighted'][0],
+                'value' => System::getFormattedNumber($this->ad['wltp_co2_emission_combined_weighted'], 1),
             ];
         }